requirements.txt
pycryptodome == 3.10.1
git+https://github.com/fionn/batch-gcd@0.0.1#egg=batch-gcd
